Refactoring server part after repository split up
Rename `server` module to `codechecker_server` and move migration
scripts under it.
