+++ alembic.ini
@@ -1,6 +1,6 @@
 [config_db]
 # path to migration scripts
-script_location = webserver/server/migrations/config
+script_location = webserver/server/codechecker_server/migrations/config
 
 # template used to generate migration files
 # file_template = %%(rev)s_%%(slug)s
@@ -31,7 +31,7 @@ sqlalchemy.url = postgres://postgres@localhost:5432/config
 
 [run_db]
 # path to migration scripts
-script_location = webserver/server/migrations/report
+script_location = webserver/server/codechecker_server/migrations/report
 
 # template used to generate migration files
 # file_template = %%(rev)s_%%(slug)s
