+++ config/package_layout.json
@@ -24,8 +24,8 @@
     "lib_tu_collector": "lib/python2.7/tu_collector",
     "libcodechecker": "lib/python2.7/libcodechecker",
     "gencodechecker": "lib/python2.7/gencodechecker",
-    "run_db_migrate": "lib/python2.7/migrations/report",
-    "config_db_migrate": "lib/python2.7/migrations/config"
+    "run_db_migrate": "lib/python2.7/codechecker_server/migrations/report",
+    "config_db_migrate": "lib/python2.7/codechecker_server/migrations/config"
   },
   "runtime": {
     "analyzers": {
