--- webserver/server/codechecker/server/database/__init__.py
@@ -1,5 +1,0 @@
-# -------------------------------------------------------------------------
-#                     The CodeChecker Infrastructure
-#   This file is distributed under the University of Illinois Open Source
-#   License. See LICENSE.TXT for details.
-# -------------------------------------------------------------------------
