+++ webserver/server/codechecker_server/api/thrift_enum_helper.py
