+++ webserver/server/codechecker_server/database/db_cleanup.py
