+++ webserver/server/codechecker_server/session_manager.py
